+++ src/utils.rs
@@ -1,6 +1,6 @@
 //  SPDX-License-Identifier: GPL-3.0-only
 /*  Build tool with support for git tags, wrapping make.
- *  Copyright (C) 2023  jgabaut
+ *  Copyright (C) 2023-2024  jgabaut
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, version 3 of the License.
@@ -36,6 +36,8 @@ pub fn print_config_args(args: &Args) {
     let execname_string: String = "E".to_owned();
     let maketag_string: String = "M".to_owned();
     let ignore_gitcheck_string: String = "X".to_owned();
+    let anvil_version_string: String = "a".to_owned();
+    let anvil_kern_string: String = "k".to_owned();
     match args.amboso_dir {
         Some(ref x) => {
             debug!("Passed amboso_dir: {{{}}}", x.display());
@@ -70,6 +72,23 @@ pub fn print_config_args(args: &Args) {
             config_string.push_str(&maketag_string);
         }
         None => {}
+    }
+    match args.anvil_version {
+        Some(ref x) => {
+            debug!("Passed anvil_version: {{{}}}", x);
+            config_string.push_str(&anvil_version_string);
+        }
+        None => {}
+    }
+    match args.anvil_kern {
+        Some(ref x) => {
+            debug!("Passed anvil_kern: {{{}}}", x);
+            config_string.push_str(&anvil_kern_string);
+        }
+        None => {
+            trace!("No anvil_kern value");
+            panic!("Missing anvil_kern value");
+        }
     }
     if args.ignore_gitcheck {
         debug!("Ignore git check is on.");